Added 3 forms on the same page for the different kinds of questions
