+++ src/main/java/org/group23/ControllerStructure.java
@@ -31,91 +31,6 @@ public class ControllerStructure {
         return "redirect:/addRemoveQuestions/" + survey.getId();
     }
 
-    @GetMapping("/addRemoveQuestions/{surveyId}")
-    public String addRemoveQuestionsForm(@PathVariable Long surveyId, Model model) {
-        // Get the survey by ID
-        Survey survey = surveyRepository.findById(surveyId).orElse(null);
-        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        String username = auth.getName();
-
-        if (survey != null && Objects.equals(username, survey.getAuthor())) {
-            model.addAttribute("survey", survey);
-            return "addRemoveQuestions";
-        } else {
-            // Handle survey not found
-            model.addAttribute(
-                    "message",
-                    "The survey was not found, or you do not have permission to edit it.");
-            return "error";
-        }
-    }
-
-    @PostMapping("/addQuestion/{surveyId}")
-    public String addQuestion(
-            @PathVariable Long surveyId,
-            @ModelAttribute("survey") Survey survey,
-            @RequestParam String questionText,
-            Model model
-    ) {
-        // Validate length before adding to the database
-        if (questionText.length() > 255) {
-            // Handle the case where the question text is too long
-            model.addAttribute("error", "Question text is too long. Please enter a shorter question.");
-            return "error";
-        }
-        // Get the survey by ID to ensure we have the latest data
-        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        String username = auth.getName();
-        Survey updatedSurvey = surveyRepository.findById(surveyId).orElse(null);
-        if (updatedSurvey != null && Objects.equals(username, updatedSurvey.getAuthor())) {
-            // Create a text question and add it to the survey
-            TextQuestion textQuestion = new TextQuestion(questionText);
-            updatedSurvey.addQuestion(textQuestion);
-            surveyRepository.save(updatedSurvey);
-
-            model.addAttribute("survey", updatedSurvey);
-
-            return "redirect:/addRemoveQuestions/" + surveyId;
-        } else {
-            // Handle survey not found
-            model.addAttribute(
-                    "message",
-                    "The survey was not found, or you do not have permission to edit it.");
-            return "error";
-        }
-    }
-
-    @PostMapping("/deleteQuestion/{surveyId}/{questionId}")
-    public String deleteQuestion(
-            @PathVariable Long surveyId,
-            @PathVariable Long questionId,
-            Model model
-    ) {
-        // Get the survey by ID
-        Survey survey = surveyRepository.findById(surveyId).orElse(null);
-        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        String username = auth.getName();
-        if (survey != null && Objects.equals(username, survey.getAuthor())) {
-            // Find and remove the question from the survey
-            Question questionToRemove = survey.getQuestions().stream()
-                    .filter(question -> question.getId().equals(questionId))
-                    .findFirst()
-                    .orElse(null);
-
-            if (questionToRemove != null) {
-                survey.removeQuestion(questionToRemove);
-                surveyRepository.save(survey);
-            }
-            return "redirect:/addRemoveQuestions/" + survey.getId();
-        } else {
-            // Handle survey not found
-            model.addAttribute(
-                    "message",
-                    "The survey was not found, or you do not have permission to edit it.");
-            return "error";
-        }
-    }
-
     @GetMapping("/saveSurvey")
     public String saveSurvey() {
         return "redirect:/surveyCreated";
